limpieza y modularizacion de componentes y formulario controlado de pasarela de pagos
