+++ ecommerce-front/src/App.jsx
@@ -26,28 +26,29 @@ function App() {
     <div>
       <Navbar />
       <Box>
-        <Routes>
-          <Route path='/' element={<Landing />} />
-          <Route path='/home' element={<Home />} />
-          <Route path='/detail/:id' element={<Details />} />
-          <Route path='/about' element={<About />} />
-         
-          <Route path='/sucursales' element={<Sucursales />} />
-          <Route path='/register' element={<FormLogin />} />
-          <Route path='/auth' element={<Auth />} />
-          <Route path='/faqs' element={<Preguntas />} />
-         
-          <Route path='/cart' element={<ShoppingCart />} />
-          <Route path='/buy' element={<Buy />}/>
-          
-          <Route path="dashboard" element={<Dashboard />} /> 
-            <Route path="users" element={<UsersPage />} >
-            // admin Routes
-
-            <Route element={<ProtectedRoute isAllowed={user && user.length>0 && user.admin?.includes('admin')} redirectTo={"/"}/>}>
-               <Route path='admin' element={<Admin />} />
-              {/* ADMIN */}
-               <Route path='/postproduct' element={<ProductForm />} />
+      <Routes>
+          <Route path="/" element={<Landing />} />
+          <Route path="/home" element={<Home />} />
+          <Route path="/detail/:id" element={<Details />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/sucursales" element={<Sucursales />} />
+          <Route path="/register" element={<FormLogin />} />
+          <Route path="/auth" element={<Auth />} />
+          <Route path="/faqs" element={<Preguntas />} />
+          <Route path="/cart" element={<ShoppingCart />} />
+          <Route path="/buy" element={<Buy />} />
+          {/* // admin Routes */}
+          <Route
+            element={
+              <ProtectedRoute
+                isAllowed={user && user.length>0 && user.admin?.includes("admin")}
+                redirectTo={"/"}
+              />
+            }
+          >
+            <Route path="admin" element={<Admin />}>
+              <Route path="dashboard" element={<Dashboard />} />
+              <Route path="users" element={<UsersPage />} />
             </Route>
             {/* ADMIN */}
             <Route path="/postproduct" element={<ProductForm />} />
@@ -60,7 +61,7 @@ function App() {
               </ProtectedRoute>
             }
           />
-        </Routes> 
+        </Routes>
       </Box>
       <Footer />
     </div>
