+++ ecommerce-front/src/components/Styled/StyledPaper.jsx
@@ -1,0 +1,15 @@
+import { styled } from "@mui/material/styles";
+import Paper from "@mui/material/Paper";
+
+export const Item = styled(Paper)(({ theme }) => ({
+    //   backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
+    backgroundColor: "none",
+    ...theme.typography.body2,
+    padding: theme.spacing(1),
+    textAlign: "center",
+    //   color: theme.palette.text.secondary,
+    color: "white",
+    background: "none",
+    border: "none",
+    boxShadow: "none",
+  }));
